Place the server closer to the center
This places the server exactly in the center of the map on the horizontal
axis for maps with odd number of columns.  Previously it was placed off
by one to the left. For example, on the "Newbie" level, it resulted in
`sourcex = 5 / 2 - 1 = 1` instead of the center column 2.

For maps with even number of columns, server placement is not changed.